requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
shap>=0.42.0
matplotlib>=3.7.0
pyarrow>=12.0.0
joblib>=1.3.0
pyyaml>=6.0
seaborn>=0.12.0
tqdm>=4.65.0
